requirements.txt
langchain-community
qdrant-client
python-dotenv
rapidfuzz
torch
transformers
